Refactored JS files
