+++ src/constants.js
@@ -1,0 +1,19 @@
+exports.col = 6;
+exports.row = 4;
+
+exports.symbolList = [
+	'http://placehold.it/100/A10559/ffffff?text=CARD',
+	'http://placehold.it/100/632065/ffffff?text=CARD',
+	'http://placehold.it/100/F9A487/ffffff?text=CARD',
+	'http://placehold.it/100/480320/ffffff?text=CARD',
+	'http://placehold.it/100/AA66FF/ffffff?text=CARD',
+	'http://placehold.it/100/BF92D5/ffffff?text=CARD',
+	'http://placehold.it/100/23AA9F/ffffff?text=CARD',
+	'http://placehold.it/100/35131E/ffffff?text=CARD',
+	'http://placehold.it/100/423A75/ffffff?text=CARD',
+	'http://placehold.it/100/F6F2E7/ffffff?text=CARD',
+	'http://placehold.it/100/C7DFA3/ffffff?text=CARD',
+	'http://placehold.it/100/FAF9E5/ffffff?text=CARD'
+];
+
+exports.DEVELOPMENT_MODE = false;
